# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/countries/": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create a country",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Countries"
                ],
                "summary": "Create a country",
                "parameters": [
                    {
                        "description": "Create a country",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CreateUpdateCountryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Country response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CountryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/countries/get-by-filter": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get Countries",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Countries"
                ],
                "summary": "Get Countries",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.PaginationInputWithFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Country response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CountryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/countries/{id}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get a country",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Countries"
                ],
                "summary": "Get a country",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Country response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CountryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Update a country",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Countries"
                ],
                "summary": "Update a country",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update a country",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CreateUpdateCountryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Country response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CountryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Delete a country",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Countries"
                ],
                "summary": "Delete a country",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_aminazadbakht1_golang-clean-web-api_api_dto.CityResponse": {
            "type": "object",
            "properties": {
                "country": {
                    "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CountryResponse"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_aminazadbakht1_golang-clean-web-api_api_dto.CompanyResponse": {
            "type": "object",
            "properties": {
                "country": {
                    "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CountryResponse"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_aminazadbakht1_golang-clean-web-api_api_dto.CountryResponse": {
            "type": "object",
            "properties": {
                "cities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CityResponse"
                    }
                },
                "companies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CompanyResponse"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_aminazadbakht1_golang-clean-web-api_api_dto.CreateUpdateCountryRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                }
            }
        },
        "github_com_aminazadbakht1_golang-clean-web-api_api_dto.Filter": {
            "type": "object",
            "properties": {
                "filterType": {
                    "description": "text number",
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "type": {
                    "description": "contains notContains equals notEqual startsWith lessThan lessThanOrEqual greaterThan greaterThanOrEqual inRange endsWith",
                    "type": "string"
                }
            }
        },
...
        "github_com_aminazadbakht1_golang-clean-web-api_api_dto.PaginationInputWithFilter": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.Filter"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "sort": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.Sort"
                    }
                }
            }
        },
...
        "github_com_aminazadbakht1_golang-clean-web-api_api_dto.Sort": {
            "type": "object",
            "properties": {
                "colId": {
                    "type": "string"
                },
                "sort": {
                    "type": "string"
                }
            }
        },
...
        "github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "result": {},
                "resultCode": {
                    "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.ResultCode"
                },
                "success": {
                    "type": "boolean"
                },
                "validationErrors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_validation.ValidationError"
                    }
                }
            }
        },
        "github_com_aminazadbakht1_golang-clean-web-api_api_helper.ResultCode": {
            "type": "integer",
            "enum": [
                0,
                40001,
                40101,
                40301,
                40401,
                42901,
                42902,
                50001,
                50002
            ],
            "x-enum-varnames": [
                "Success",
                "ValidationError",
                "AuthError",
                "ForbiddenError",
                "NotFoundError",
                "LimiterError",
                "OtpLimiterError",
                "CustomRecovery",
                "InternalError"
            ]
        },
        "github_com_aminazadbakht1_golang-clean-web-api_api_validation.ValidationError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "property": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AuthBearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}